requirements.txt
Pillow
Flask
pylint
cairosvg
